[Editorial] remove square brackets from chapter 1
The text is littered with things in square brackets that have been added in by
the translator. They often provide additional context, or else they're used to
clarify subjects that are ambiguous, while (presumably) leaving the literal
translation intact.

However, formatting the text in this way is very distracting for casual
readers, so where possible I've either removed the brackets or removed the
interjections entirely.

Where they're adding context, but the meaning of the text is otherwise clear,
I've moved them to endnotes.

I've kept a few of the brackets in. I think the ones surrounded by quotation
marks are fine, although the Chicago Manual of Style says to use them
sparingly (https://www.chicagomanualofstyle.org/book/ed17/part2/ch13/psec060.html).

The original text comes with extensive footnotes, but these aren't present in
the Project Gutenberg source, so I haven't added them in yet. They don't always
add much to the text, but a lot of them provide useful historical and geographic
context.

Often the text will mention such and such's consulship, and actual dates are
provided both in brackets and footnotes. Ideally I'd like to provide all of
these as endnotes.

In a couple of places Roman miles are convered to "english miles". I don't
think this is particularly useful as this is simply another historical
measurement that readers are unlikely to be familiar with, so I've left the
text saying "miles" and clarified that these are roman miles in the endnotes.

The 4th endnote paraphrases the translator's footnote. Originally "according to
their own representations" was included in square brackets, and the footnote
explained why it's in square brackets. As it's not essential for understanding
the text, I'd prefer to use a single endnote instead.

I've changed "seventh hour" and "eventide" to "noon" and "evening" to
remove the archaic term and the need for clarifying brackets.