[Editorial] Rework changes to square brackets
This is the logic I've tried to follow when deciding how to deal with these:

1. If the text in brackets adds historical/geographical context that wasn't
obvious from the surrounding text, keep it. Longer examples can be moved to
endnotes, but I've left them inline if they don't relate to the main point of
the sentence.

2. If it's not possible to omit the brackets without modifying the sentence
structure, keep them. If the brackets clarify who/what is being talked about
keep them as square brackets. If they are used for interjections like "on the
one hand", change them to parentheses.

3. Otherwise, remove the square brackets